exoself (nn v1.0.0)

Functions

map()

-spec map() -> pid().

Map genotype to phenotype (default filename: ffnn)

Convenience function that loads the default genotype file and spawns an ExoSelf process to train it.

Returns

PID of the spawned ExoSelf process

Examples

  exoself:map().

map(FileName)

-spec map(atom()) -> pid().

Map genotype to phenotype

Loads a neural network genotype from file and spawns an ExoSelf process to construct the phenotype and begin training.

This is the main entry point for running a neural network.

Parameters

- FileName - Name of the genotype file (atom)

Returns

PID of the spawned ExoSelf process

Examples

  % Create a genotype first
  genotype:construct(my_network, xor_mimic, [3]).
 
  % Map it to a running phenotype
  exoself:map(my_network).