morphology (nn v1.0.0)

Morphology Definitions for Neural Networks

This module defines different neural network morphologies - configurations of sensors and actuators for specific problem domains. Each morphology specifies the input/output interface for a particular task.
A morphology defines:

Sensors: Input components that receive data from the environment

Actuators: Output components that send actions to the environment

Scapes: Private or public environment simulators

Currently implemented morphologies:

xor_mimic - XOR problem with 2 inputs and 1 output

Types

actuator_id/0

-type actuator_id() :: {component_id(), actuator}.

component_id/0

-type component_id() :: {layer_index(), unique_id()}.

cortex_id/0

-type cortex_id() :: {component_id(), cortex}.

layer_index/0

-type layer_index() :: integer().

morphology/0

-type morphology() :: atom().

neuron_id/0

-type neuron_id() :: {component_id(), neuron}.

scape_name/0

-type scape_name() :: {private | public, atom()}.

sensor_id/0

-type sensor_id() :: {component_id(), sensor}.

unique_id/0

-type unique_id() :: float().

Functions

get_Actuators(Morphology)

-spec get_Actuators(Morphology :: morphology()) ->
                       [#actuator{id :: actuator_id(),
                                  name :: atom(),
                                  cortex_id :: cortex_id(),
                                  scape :: scape_name(),
                                  vector_length :: pos_integer(),
                                  fanin_ids :: [neuron_id()],
                                  generation :: non_neg_integer()}].

Alias for get_InitActuators/1
See also: get_InitActuators/1.

get_InitActuator(Morphology)

-spec get_InitActuator(Morphology :: morphology()) ->
                          #actuator{id :: actuator_id(),
                                    name :: atom(),
                                    cortex_id :: cortex_id(),
                                    scape :: scape_name(),
                                    vector_length :: pos_integer(),
                                    fanin_ids :: [neuron_id()],
                                    generation :: non_neg_integer()}.

Get the first actuator for a morphology
Retrieves the first actuator from the list of actuators defined for the specified morphology. Useful when only one actuator is needed.
See also: get_InitActuators/1.

get_InitActuators(Morphology)

-spec get_InitActuators(Morphology :: morphology()) ->
                           [#actuator{id :: actuator_id(),
                                      name :: atom(),
                                      cortex_id :: cortex_id(),
                                      scape :: scape_name(),
                                      vector_length :: pos_integer(),
                                      fanin_ids :: [neuron_id()],
                                      generation :: non_neg_integer()}].

Get all actuators for a morphology
Retrieves the complete list of actuators defined for the specified morphology. Each actuator is pre-configured with its scape, vector length, and function name.
See also: get_InitActuator/1.

get_InitSensor(Morphology)

-spec get_InitSensor(Morphology :: morphology()) ->
                        #sensor{id :: sensor_id(),
                                name :: atom(),
                                cortex_id :: cortex_id(),
                                scape :: scape_name(),
                                vector_length :: pos_integer(),
                                fanout_ids :: [neuron_id()],
                                generation :: non_neg_integer()}.

Get the first sensor for a morphology
Retrieves the first sensor from the list of sensors defined for the specified morphology. Useful when only one sensor is needed.
See also: get_InitSensors/1.

get_InitSensors(Morphology)

-spec get_InitSensors(Morphology :: morphology()) ->
                         [#sensor{id :: sensor_id(),
                                  name :: atom(),
                                  cortex_id :: cortex_id(),
                                  scape :: scape_name(),
                                  vector_length :: pos_integer(),
                                  fanout_ids :: [neuron_id()],
                                  generation :: non_neg_integer()}].

Get all sensors for a morphology
Retrieves the complete list of sensors defined for the specified morphology. Each sensor is pre-configured with its scape, vector length, and function name.
See also: get_InitSensor/1.

get_Sensors(Morphology)

-spec get_Sensors(Morphology :: morphology()) ->
                     [#sensor{id :: sensor_id(),
                              name :: atom(),
                              cortex_id :: cortex_id(),
                              scape :: scape_name(),
                              vector_length :: pos_integer(),
                              fanout_ids :: [neuron_id()],
                              generation :: non_neg_integer()}].

Alias for get_InitSensors/1
See also: get_InitSensors/1.

xor_mimic(Type)

-spec xor_mimic(Type :: sensors | actuators) ->
                   [#sensor{id :: sensor_id(),
                            name :: atom(),
                            cortex_id :: cortex_id(),
                            scape :: scape_name(),
                            vector_length :: pos_integer(),
                            fanout_ids :: [neuron_id()],
                            generation :: non_neg_integer()}] |
                   [#actuator{id :: actuator_id(),
                              name :: atom(),
                              cortex_id :: cortex_id(),
                              scape :: scape_name(),
                              vector_length :: pos_integer(),
                              fanin_ids :: [neuron_id()],
                              generation :: non_neg_integer()}].

XOR Mimic Morphology

Defines a morphology for solving the XOR (exclusive OR) problem.
Problem Description: XOR is a classic non-linearly separable problem that requires at least one hidden layer to solve. The truth table is:
  Input1  Input2  Output
    -1      -1      -1
     1      -1       1
    -1       1       1
     1       1      -1
Sensors:

Name: xor_GetInput

Vector Length: 2 (two binary inputs)

Scape: {private, xor_sim} - each agent gets its own simulator

Actuators:

Name: xor_SendOutput

Vector Length: 1 (single binary output)

Scape: {private, xor_sim} - same simulator as sensor